\name{DisentangleSNP-package}
\alias{DisentangleSNP-package}
\alias{DisentangleSNP-pkg}
\docType{package}
\title{
\packageTitle{DisentangleSNP}
}
\description{
\packageDescription{DisentangleSNP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DisentangleSNP}
\packageIndices{DisentangleSNP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DisentangleSNP}

Maintainer: \packageMaintainer{DisentangleSNP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
